Revert testutil move
