+++ lib/testhelper.js
@@ -12,7 +12,7 @@ governing permissions and limitations under the License.
 
 'use strict';
 
-const sendQueue = require("../lib/queue");
+const sendQueue = require("./queue");
 const zlib = require('zlib');
 const sleep = require('util').promisify(setTimeout);
 const assert = require("assert");
